requirements.txt
# Iris dependencies
# Install with: pip3 install --user -r requirements.txt

# Spreadsheet parsing
openpyxl>=3.1

# Camera preview/capture UI uses PIL Image/ImageTk
Pillow>=10.0

# Use opencv-python on desktop/dev.
# On Jetson, prefer apt package (python3-opencv) for GStreamer support.
opencv-python>=4.8; platform_system != "Linux" or platform_machine != "aarch64"
